refactor(Server): remove server.go and web-sockets.go files, consolidating functionality and improving code organization
