+++ beam/server/handlers/files.go
@@ -121,4 +121,3 @@ func sendJSONError(w http.ResponseWriter, message string, statusCode int) {
 	w.WriteHeader(statusCode)
 	json.NewEncoder(w).Encode(map[string]string{"error": message})
 }
-
