+++ beam/types.go
@@ -8,3 +8,11 @@ type ServerStats struct{
     Uploads   int       `json:"uploads"`
     StartTime time.Time `json:"startTime"`
 }
+
+type File struct {
+	Name    string `json:"name"`
+	Size    string `json:"size"`
+	IsDir   bool   `json:"isDir"`
+	ModTime string `json:"modTime"`
+	Path    string `json:"path"`
+}
